+++ src/Services/ModalService.ts
@@ -184,7 +184,15 @@ module StreamStats.Services {
                             backdrop: 'static',
                             windowClass: 'windowZ',
                         };
-     
+                    case SSModalType.e_warningmessage:
+                        return {
+                            templateUrl: 'Views/warningmessage.html',
+                            controller: 'StreamStats.Controllers.WarningMessageController',
+                            size: 'sm',
+                            backdropClass: 'backdropZ',
+                            backdrop: 'static',
+                            windowClass: 'windowZ',
+                        };
                     default:
                         return null;
                 }//end switch          
@@ -208,7 +216,8 @@ export enum SSModalType {
         e_extensionsupport = 9,
         e_gagepage = 10,
         e_nearestgages = 11,        
-        e_flowanywhere = 12
+        e_flowanywhere = 12,
+        e_warningmessage = 13
         
         
     }